Add option to log test-case and step output
Even if you are running a test case or step on its own, if you
pass `-l` or `--log_everything` to `compass run`, the output from
the test case or step will go to a log file in the current
(test case or step) work directory.  This will be useful for
situations where steps are run as separate subprocesses as
part of task parallelism.