Switch from run_model to add_model_substeps
This merge adds two classes for substeps needed to run the MPAS
model.  The first does setup such as updating namelists for PIO
and using Metis to create a graph partition file for the desired
number of tasks.  The second runs the model itself.

A new function `add_model_substeps()` adds these substeps to a
step.