Optionally, run only one substep of a step
When calling `compass run` in a step directory, optionally
pass the `--substep=<substep>` flag to indicate that a specific
substep should be run.  This should be useful in the future for
task parallelism because it will allow substeps to be run
individually with calls to `srun` with a specified number of
resources.